requirements.txt
torch
bitsandbytes
accelerate
transformers
huggingface-hub
datasets
sentencepiece
safetensors
langchain
langchain-huggingface
Pillow
opencv-python
requests
tqdm
ipython
flask
tiktoken
gunicorn
